requirements.txt
openai
requests
python-dotenv
Flask